0.0.02.001
remove legacy track length limit; max track length is now range of signed 32-bit integer (2,147,483,647)

+++ trunk/Polymeter/TrackView.cpp
@@ -160,7 +160,7 @@ CWnd *CTrackView::CTrackGridCtrl::CreateEditCtrl(LPCTSTR pszText, DWORD dwStyle,
 			pEdit->SetRange(1, SHRT_MAX);
 			break;
 		case COL_Length:
-			pEdit->SetRange(1, SHRT_MAX);
+			pEdit->SetRange(1, INT_MAX);
 			break;
 		case COL_Velocity:
 			pEdit->SetRange(-127, 127);
